Move road connection functions to `Connections.cpp`
