+++ appOPHD/Map/Connections.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <string>
+
+
+class TileMap;
+class Road;
+
+
+std::string roadAnimationName(const Road& road, const TileMap& tileMap);
